linkerror2005修正
把definition.h里的类外函数搬到functions_definition.cpp即可
