+++ local_logic/local_logic/definition.h
@@ -103,20 +103,10 @@ struct TPoint
 };
 
 
-TPoint operator-(const TPoint& p1, const TPoint& p2)
-{
-	TPoint temp;
-	temp.m_x = p1.m_x - p2.m_x;
-	temp.m_y = p1.m_y - p2.m_y;
-	temp.m_state = Disables;
-	return temp;
-}
+TPoint operator-(const TPoint& p1, const TPoint& p2);
 
 //计算欧式距离
-TLength getDistance(const TPoint& p1, const TPoint& p2)
-{
-	return sqrt(pow(p1.m_x - p2.m_x, 2) + pow(p1.m_y - p2.m_y, 2));
-}
+TLength getDistance(const TPoint& p1, const TPoint& p2);
 
 enum StudentType  //学生种类的枚举
 {
